fix(product): refuse to confirm a shelf the holds already outrun
Saying yes to the recheck locks the number in. If the holds are already
above it, that locks in a shortage the shop then has to wait out, so the
answer is refused until the overflow is dealt with.

The rest is housekeeping the split left behind. dropHeldQty was
releaseHold under another name, with a comment claiming it left the
available quantity alone while the line below recomputed it.
takeFromAvailable checks the available quantity and subtracts from the
shelf, so it is takeFromStock. The threshold that decides how far a
quantity may fall lived as a double between two methods and as a number
inside an error message; it is one pair of integer constants now, and the
message no longer repeats it. adjustAvailableQty is private, so the
derived quantity can only come from stock and holds.

+++ src/main/java/com/swyp/backend/home/service/OwnerHomeService.java
@@ -48,7 +48,7 @@ public class OwnerHomeService {
 				holdFunction.countCompletedTodayOfStore(storeId),
 				sumAvailableQty(productCards));
 		OwnerHomeIssues issues = new OwnerHomeIssues(
-				holdFunction.countExpiredTodayOfStore(storeId),
+				(int) holdFunction.countExpiredTodayOfStore(storeId),
 				(int) products.stream().filter(product -> product.shortfallQty() > 0).count(),
 				products.stream().mapToInt(Product::shortfallQty).sum());
 		long unreadNotificationCount = notificationFunction.countUnread(ownerId);
